Move more sign-in related stuff into React land
