+++ src/hub.js
@@ -689,72 +689,6 @@ document.addEventListener("DOMContentLoaded", async () => {
     });
   };
 
-  window.addEventListener("action_create_avatar", () => {
-    performConditionalSignIn(
-      () => hubChannel.signedIn,
-      () => pushHistoryState(history, "overlay", "avatar-editor"),
-      "create-avatar"
-    );
-  });
-
-  scene.addEventListener("scene_media_selected", e => {
-    const sceneInfo = e.detail;
-
-    performConditionalSignIn(
-      () => hubChannel.can("update_hub"),
-      () => hubChannel.updateScene(sceneInfo),
-      "change-scene"
-    );
-  });
-
-  scene.addEventListener("action_media_tweet", e => {
-    let isInModal = false;
-    let isInOAuth = false;
-
-    const exitOAuth = () => {
-      isInOAuth = false;
-      store.clearOnLoadActions();
-      remountUI({ showOAuthDialog: false, oauthInfo: null });
-    };
-
-    handleExitTo2DInterstitial(true, () => {
-      if (isInModal) history.goBack();
-      if (isInOAuth) exitOAuth();
-    });
-
-    performConditionalSignIn(
-      () => hubChannel.signedIn,
-      async () => {
-        // Strip el from stored payload because it won't serialize into the store.
-        const serializableDetail = {};
-        Object.assign(serializableDetail, e.detail);
-        delete serializableDetail.el;
-
-        if (hubChannel.can("tweet")) {
-          isInModal = true;
-          pushHistoryState(history, "modal", "tweet", serializableDetail);
-        } else {
-          if (e.detail.el) {
-            // Pin the object if we have to go through OAuth, since the page will refresh and
-            // the object will otherwise be removed
-            e.detail.el.setAttribute("pinnable", "pinned", true);
-          }
-
-          const url = await hubChannel.getTwitterOAuthURL();
-
-          isInOAuth = true;
-          store.enqueueOnLoadAction("emit_scene_event", { event: "action_media_tweet", detail: serializableDetail });
-          remountUI({
-            showOAuthDialog: true,
-            oauthInfo: [{ type: "twitter", url: url }],
-            onCloseOAuthDialog: () => exitOAuth()
-          });
-        }
-      },
-      "tweet"
-    );
-  });
-
   remountUI({ performConditionalSignIn, embed: isEmbed, showPreload: isEmbed });
   entryManager.performConditionalSignIn = performConditionalSignIn;
   entryManager.init();
